Using the parent json in the same folder instead.
The json file doesn't content any element but there is a parent so this will find the parent inside the same folder directory.
There is no json file from the parent it will display an error to tell there no acceptable json file. Finding a json manually. 
Sorry I'm new to this and my English is horrible.